Create Schedule and User

+++ src/main/java/com/example/todo/service/UserService.java
@@ -18,9 +18,9 @@ public class UserService {
 
     private final UserRepository userRepository;
 
-    public UserResponseDto save(String username, String email) {
+    public UserResponseDto save(String username, String email, String password) {
 
-        User user = new User(username, email);
+        User user = new User(username, email, password);
 
         User saveUser = userRepository.save(user);//값이 알아서 속성에 맞게 매핑되어있음
 
@@ -30,28 +30,36 @@ public class UserService {
 
     public UserResponseDto findById(Long id) {
 
-        Optional<User> optionalUser = userRepository.findById(id);
-
-        if (optionalUser.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-        }
-
-        User findUser = optionalUser.get();//optional에 내장된 매서드 get()은 optional에 저장된 객체(값)을 꺼낸다.
+        User findUser = userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST));
+        //optional에 내장된 매서드 get()은 optional에 저장된 객체(값)을 꺼낸다.
 
         return new UserResponseDto(findUser.getId(), findUser.getUsername(), findUser.getEmail());
     }
 
 
-    public UserResponseDto updateUser(Long id, String username, String email) {
+    public UserResponseDto updateUser(Long id, String password, String username, String email) {
+        User findUser = userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST));
+//데이터베이스에서 데이터 가져올 때 무조건 OPTIONAL이나 예외 처리를 해야하는게 규칙인가?
+// id를 넘거야 어떤 엔티티를 수정할지 식별할 수 있음
 
-        User user = new User(id, username, email);//id를 넘거야 어떤 엔티티를 수정할지 식별할 수 있음
+        if (findUser.getPassword().equals(password)) {
+            //JPA에서는 SAVE를 사용하면 자동으로 INSERT나 UPDATE를 수행한다.
+            findUser.editUser(username, email);//user 클래스 값을 매서드로 바꿔줌
 
-        User updateUser = userRepository.save(user);//JPA에서는 SAVE를 사용하면 자동으로 INSERT나 UPDATE를 수행한다.
+            return new UserResponseDto(findUser.getId(), findUser.getUsername(), findUser.getEmail());
 
-        return new UserResponseDto(updateUser.getId(), updateUser.getUsername(), updateUser.getEmail());
+        } else throw new IllegalArgumentException("비밀번호가 다릅니다");
     }
 
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+    public void deleteUser(Long id, String password) {
+        User findUser = userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST));
+        if (findUser.getPassword().equals(password)) {
+
+            userRepository.deleteById(id);
+        }
+        else throw new IllegalArgumentException("비밀번호가 다릅니다");
+
+
+
     }
 }
